Many small changes to remove warnings.

+++ include/utilities_slsim.h
@@ -1288,13 +1288,13 @@ void apply_permutation(
 
 /** \brief Calculates power spectrum from a 2d map or the cross-power spectrum between two 2d maps.
  *
- *   Adaptied from Carlo Giocoli's pl() routine.
+ *   Adapted from Carlo Giocoli's pl() routine.
  */
 void powerspectrum2d(
                      std::valarray<double> const &aa      /// first realspace map to be
                      ,std::valarray<double> const &bb     /// second realspace map, same as aa to get power spectrum
-                     ,int nx                       /// number of pixels in x direction
-                     ,int ny                       /// number of pixels in y direction
+                     ,long nx                       /// number of pixels in x direction
+                     ,long ny                       /// number of pixels in y direction
                      ,double boxlx                 /// range of image in x direction
                      ,double boxly                 /// range of image in y direction
                      ,std::vector<double> &ll      /// output multiplot number of bins
@@ -1304,8 +1304,8 @@ void powerspectrum2d(
 
 void powerspectrum2d(
                      std::valarray<double> &aa      /// first realspace map to be
-                     ,int nx                       /// number of pixels in x direction
-                     ,int ny                       /// number of pixels in y direction
+                     ,long nx                       /// number of pixels in x direction
+                     ,long ny                       /// number of pixels in y direction
                      ,double boxlx                 /// range of image in x direction
                      ,double boxly                 /// range of image in y direction
                      ,std::vector<double> &ll      /// output multiplot number of bins
@@ -1314,8 +1314,8 @@ void powerspectrum2d(
 
 void powerspectrum2dprebin(
                      std::valarray<double> &aa      /// first realspace map to be
-                     ,int nx                       /// number of pixels in x direction
-                     ,int ny                       /// number of pixels in y direction
+                     ,long nx                       /// number of pixels in x direction
+                     ,long ny                       /// number of pixels in y direction
                      ,double boxlx                 /// range of image in x direction
                      ,double boxly                 /// range of image in y direction
                      ,const std::vector<double> &ll      /// output multiplot number of bins
